Converting offer 'update_ts' fields to datetime objects.
This adds the third party python-dateutil (http://labix.org/python-dateutil) module as a dependancy.
This is necessary because the Octopart API gives timestamps as ISO 8601-formatted strings.
While datetime.strptime() can parse most of ISO 8601, it does not understand the use of "Z" to denote UTC.

Other possible solutions ivestigated (all third-party modules):
-Use the pytz Olson timezone database module (http://pytz.sourceforge.net/) with strptime():
This would not work because the Olson short name for Zulu time is "UTC", not "Z".

-Parsing with pyiso8601 (http://code.google.com/p/pyiso8601/) instead of python-dateutil:
In theory, this should work just as well.
The deciding factor was that pyiso8601 was last updated in 2007, while python-dateutil is being actively maintained.

Signed-off-by: Joe Baker <jbaker@alum.wpi.edu>